Extract initSettlement to own class
